requirements.txt
pillow
matplotlib
python-dotenv

# - e